refactor(internal): drop pkg layer
