+++ internal/assets/.arikawa-boilerplate.example.yaml
