+++ internal/middleware/ratelimitable.go
@@ -7,8 +7,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/Serpentiel/arikawa-boilerplate/internal/container"
 	"github.com/Serpentiel/arikawa-boilerplate/internal/pkg/builder"
-	"github.com/Serpentiel/arikawa-boilerplate/internal/pkg/container"
 	"github.com/Serpentiel/arikawa-boilerplate/pkg/logger"
 	"github.com/diamondburned/arikawa/v3/api"
 	"github.com/diamondburned/arikawa/v3/api/cmdroute"
