+++ internal/pkg/builder/messageresponse.go
@@ -8,7 +8,7 @@ import (
 	"math/rand"
 	"net/http"
 
-	"github.com/Serpentiel/arikawa-boilerplate/internal/pkg/container"
+	"github.com/Serpentiel/arikawa-boilerplate/internal/container"
 	"github.com/Serpentiel/arikawa-boilerplate/pkg/logger"
 	"github.com/diamondburned/arikawa/v3/api"
 	"github.com/diamondburned/arikawa/v3/discord"
