+++ internal/provide/manager.go
@@ -5,10 +5,10 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/Serpentiel/arikawa-boilerplate/internal/pkg/cmd"
-	"github.com/Serpentiel/arikawa-boilerplate/internal/pkg/container"
-	"github.com/Serpentiel/arikawa-boilerplate/internal/pkg/presence"
-	"github.com/Serpentiel/arikawa-boilerplate/internal/pkg/router"
+	"github.com/Serpentiel/arikawa-boilerplate/internal/cmd"
+	"github.com/Serpentiel/arikawa-boilerplate/internal/container"
+	"github.com/Serpentiel/arikawa-boilerplate/internal/presence"
+	"github.com/Serpentiel/arikawa-boilerplate/internal/router"
 	"github.com/Serpentiel/arikawa-boilerplate/pkg/logger"
 	"github.com/diamondburned/arikawa/v3/gateway"
 	"github.com/diamondburned/arikawa/v3/session/shard"
