+++ internal/router/router.go
@@ -4,9 +4,9 @@ package router
 import (
 	"net/http"
 
-	"github.com/Serpentiel/arikawa-boilerplate/internal/pkg/cmd"
-	"github.com/Serpentiel/arikawa-boilerplate/internal/pkg/container"
-	"github.com/Serpentiel/arikawa-boilerplate/internal/pkg/middleware"
+	"github.com/Serpentiel/arikawa-boilerplate/internal/cmd"
+	"github.com/Serpentiel/arikawa-boilerplate/internal/container"
+	"github.com/Serpentiel/arikawa-boilerplate/internal/middleware"
 	"github.com/Serpentiel/arikawa-boilerplate/pkg/logger"
 	"github.com/diamondburned/arikawa/v3/api"
 	"github.com/diamondburned/arikawa/v3/api/cmdroute"
